requirements.txt
beautifulsoup4==4.12.3
requests==2.31.0
progressbar2==4.3.2